flask/requirements.txt
opencv-python
numpy
matplotlib
pdf2image
pytesseract
pillow
PyPDF2 
selenium
requests
webdriver-manager
flask